feat: movies block
