+++ src/components/MoviesCardList/MoviesCardList.css
@@ -26,63 +26,3 @@
 .cards__button:hover {
     filter: invert();
 }
-
-.card {
-    width: 100%;
-    background: #FDFDFD;
-    box-shadow: 0px 4px 30px rgba(0, 0, 0, 0.08);
-    transition: .2s;
-    cursor: pointer;
-}
-
-.card:hover {
-    transform: scale(1.05);
-}
-
-.card__box {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin: 14px;
-}
-
-.card__title {
-    font-size: 15px;
-    line-height: 18px;
-    margin: 0 0 9px 0;
-}
-
-.card__button {
-    width: 30px;
-    height: 30px;
-    border: none;
-    border-radius: 50%;
-    background-image: url("../../assets//images/save.svg");
-    background-repeat: no-repeat;
-    background-size: 30%;
-    background-position: center;
-    background-color: #E8E8E8;
-    cursor: pointer;
-    transition: .2s;
-}
-
-.card__button:hover {
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.5);
-}
-
-.card__button_saved {
-    background-image: url("../../assets//images/save.svg");
-    background-color: black;
-}
-
-.card__subtitle {
-    font-size: 11px;
-    line-height: 13px;
-    color: #A0A0A0;
-}
-
-.card__image {
-    width: 100%;
-    height: 203px;
-    object-fit: cover;
-}
