+++ src/components/Portfolio/Portfolio.css
@@ -1,0 +1,60 @@
+.portfolio {
+    list-style-type: none;
+    padding: 0;
+    margin: 0 70px 125px;
+
+}
+
+.portfolio__list-item {
+    width: 100%;
+}
+
+.portfolio__list-item:nth-of-type(2) {
+    border-top: 1px solid #E8E8E8;
+    border-bottom: 1px solid #E8E8E8;
+}
+
+.portfolio__link {
+    display: flex;
+    color: black;
+    text-decoration: none;
+    font-weight: normal;
+    font-size: 30px;
+    line-height: 50px;
+    letter-spacing: -0.04em;
+    width: 100%;
+    height: 100%;
+    padding: 20px 0;
+    background-image: url("../../assets/images/arrow.svg");
+    background-repeat: no-repeat;
+    background-size: 60px 29px;
+    background-position-x: right;
+    background-position-y: center;
+    transition: 0.2s;
+}
+
+.portfolio__link:hover {
+    background-color: #FD5602;
+    padding-left: 20px;
+}
+
+@media screen and (max-width: 768px) {
+    .portfolio__link {
+        font-size: 28px;
+        line-height: 50px;
+    }
+}
+
+@media screen and (max-width: 550px) {
+    .portfolio {
+        list-style-type: none;
+        padding: 0;
+        margin: 0;
+    }
+
+    .portfolio__link {
+        font-size: 18px;
+        line-height: 28px;
+        background-size: 15px 25px;
+    }
+}
